fix parameter for client also, fix cast conversion

+++ ROS/catkin_ws/src/remote_control/scripts/remoteControlServer.py
@@ -64,7 +64,7 @@ def saveNewConfigs(currentConfigs, additionalArgs):
 		elif key == "--serialPath":
 			currentConfigs["portPath"] = value
 		elif key == "--serialTimeout":
-			currentConfigs["serialTimeout"] = int(value)
+			currentConfigs["serialTimeout"] = float(value)
 		elif key == "--serialBaudRate":
 			currentConfigs["serialBaudRate"] = int(value)
 		elif key == "--maxLen":
